Fix: fetch the count instead of fetching all the receipts

+++ aggregation_mode/batcher/src/config.rs
@@ -8,7 +8,7 @@ pub struct Config {
     pub db_connection_url: String,
     pub eth_rpc_url: String,
     pub payment_service_address: String,
-    pub max_proofs_per_day: usize,
+    pub max_proofs_per_day: i64,
 }
 
 impl Config {
